+++ src/components/Home/index.js
@@ -1,0 +1,3 @@
+export { Category } from "../components/Home/Category";
+export { Hero } from "../components/Home/Hero";
+export { Trending } from "../components/Home/Trending";
