+++ src/components/Navbar/Navbar.js
@@ -4,13 +4,12 @@ import SideNav from "react-simple-sidenav";
 import "../../App.css";
 import "./Navbar.css";
 import { titleStyle, itemStyle } from "./sidebar-style";
-import { useWishlist } from "../../context/wishlist-context";
-import { useCart } from "../../context/cart-context";
+import { useCart, useWishlist } from "../../context";
 
 export const Navbar = () => {
   const [showSideNav, setShowSideNav] = useState(false);
-  const {wishlistState} = useWishlist();
-  const {cartState} = useCart();
+  const { wishlistState } = useWishlist();
+  const { cartState } = useCart();
 
   const routeData = [
     {
@@ -33,11 +32,7 @@ export const Navbar = () => {
         "Wishlist"
       ) : (
         <i class="far fa-bookmark icon-link-3pt">
-          
-          <span className="badge">
-          {wishlistState.length}
-
-          </span>
+          <span className="badge">{wishlistState.length}</span>
         </i>
       ),
     },
@@ -48,10 +43,7 @@ export const Navbar = () => {
         "Cart"
       ) : (
         <i class="fas fa-shopping-cart icon-link-3pt">
-          <span className="badge">
-          {cartState.length}
-
-          </span>
+          <span className="badge">{cartState.length}</span>
         </i>
       ),
     },
@@ -110,7 +102,6 @@ export const Navbar = () => {
           <input type="search" class="search" placeholder="Search.." />
         </div>
       </header>
-      
     </>
   );
 };
