+++ src/context/index.js
@@ -1,0 +1,3 @@
+export { useCart } from "./cart-context";
+export { useFilter } from "./filter-context";
+export { useWishlist } from "./wishlist-context";
