+++ src/pages/Home.js
@@ -1,7 +1,5 @@
 import React from "react";
-import { Category } from "../components/Home/Category";
-import { Hero } from "../components/Home/Hero";
-import { Trending } from "../components/Home/Trending";
+import { Category, Hero, Trending } from "../components/Home";
 
 export const Home = () => {
   return (
